Added "Use Term" button in term info that commits the currently browsed term to the datamodel - this was sorely needed for a while now as it was silly to browse a term and then not be able to shove it in. eventually would like to be able to do with a double click - need to somehow supress first click.
TermInfo sends out update transaction with obo class - CharFieldEnum
is figured via OnotologyManager - figures out which field needs to get
updated - and encodes in transaction.

CharFieldGui listens for char change event which maybe should be
called char field change event - and queries if the field is its field
and if so updates to string in event, and thus synchs with model.