initial config files are now copied to .phenote/my-phenote.cfg file. this allows the config file to be accessible when running off webstart jar - previoulsy config files were buried in jar in accessible.
theres now 2 cmd line options -i & -c - -i is for initial config and
will only be used if nothing in my-phenote.cfg - and if so will get
copied to my-phenote - -c overrides and overwrites whatevers in
my-phenote.

eventually there will be a gui for config as users dont want to have
to mutz with a xml file.