well this is bit of a hack that i want to revisit - so now syns are being queried from web - and syn is displayed not term name in auto comp list - and when you select the syn its term name shows up in the field - my first attempt at doing this was to just add a onmouseclick function to the li list terms - the function gets called but then control.js does its thing and sets with syn - so now i have the onmouseclick function in the li (from backend servlet) call selectterm javascript function in ncbo.html. this function sets the elements attribute/property termName to the the termName, and then i modified controls.js updateElement function to check for the presence of termName.
so this works but its hacky and i would like to revisit.

so i think maybe a better way of transmitting to the term name -
though still not great - is perhaps through the li elements id (or
title?) - sohel had been setting the id to the id of the term and i
followed suit - i dont think any of ncbo/zfin stuff uses this so maybe
its ok to make that the term name - but mayb dichty uses the id???
sohel?
if so could use title? (a bit cheesy)
then controls.js could be subclassed instead of modified in place -
and the subclass would use the id or title

or the subclass could call the onmouseclick function itself - though
im not sure if thats possible - dont know if the param is available at
that point

well im open to suggestions from javascript fiends