so added new queryable data adapter interface - char field gui queries if there is a queryable data adapter - typically its a database adapter - and if so queries as to if its char field is queryable and if so puts in a retrieve button which when pressed send field name and user input to queryable data adapter for it to query on and return a new CharacterListI. configured with <queryable-dataadapter name... enabled... />
made a stub of a worm adapter - doesnt do any querying yet (juancarlos...) but
it says that it can query on Allele and Pub (gets strings from CharFieldEnum.
thus the allele field and pub field show retrieve buttons that just bring up
at the moment a not yet implemented message.

throws DataAdapterEx if fails to retrieve as it does at the moment

woops need to also add the preliminary worm.cfg