i was getting an exception when multi selecting rows that had lists that i really had to dig for - but i think ive got it - and i just wanna make sure it sounds right.
so on multi select in AbstractAutoCompList setValsFromChars is called that then calls setMultiValueConditions -> setGuiForMultiValue which sets text to "Multiple Values" (if values are different)
and sets flag settingTextExternally to true - which says the user isnt setting this text its getting set by some external process - crucial
setText triggers doCompletion

then in doCompletion i found:


private void doCompletion(boolean showPopup, boolean thread) {
  this.setHasChangedMultipleValues(true);
   if (settingTextExternally) // flag set if text filled in externally (from table sel)
      return;


and i think this is backwards - that we dont want to set the sethaschangedmultiplevalues flag if we are stting the text externally - right? as we are setting it to "Multiple Values" the multi select text - though im not sure about selection setting text externally(?) and thus the question

but if i switch these lines and return on settextExternally before seting HCMV this bug goes away
as otherwise HasChangedMv was set so then in ComboBoxActionListener when focus was lost:

     if (AbstractAutoCompList.this.shouldResetGuiForMultipleValues()) {
       AbstractAutoCompList.this.setGuiForMultipleValues();
       return;
     } else {
       log().debug("Action updating model: " + e);
       AbstractAutoCompList.this.updateModel();
     }

shouldResetGFMV was returning false when it should be true on multi select and thus is was going ahead with updateModel below which was causing a null value to be added to the list as there wasnt a valid value (which maybe i should catch anyways just in case? as it should never happen) and the null list value was causing null pointer exception (on checking for dups yada yada)

so jim - is there any reason it might be problematic to switch those lines - is there a reason they were in that order - if so we may need more logic in here so the setting to "Multiple Values" doesnt trigger a completion as it shouldnt