Tested a new stored procedure that will - do away with stack overflow problems on the JVM while returning phenotype summaries for higher taxa and anatomical entities, and - return accurate counts for publications, genes, and taxa while accounting for subsumptions among the entities (case 667 on FogBugz)
At present, this procedure is slow to execute taking nearly 90 seconds to find the phenotype summary for Siluriformes. The advantage is that it returns about 1900 rows that can be channeled into the JSON object on the client with very little post query processinf. Compare this with the current procedure where nearly 390,000 rows will be returned for Siluriformes and fed into DTOs (which causes the memory to overflow) before being processed to generate the phenotype summary.